requirements.txt
google-adk
fastapi
python-telegram-bot
requests
pydantic
python-dotenv
litellm